pyproject.toml
[project]
name = "arch-sparring-agent"
version = "0.1.0"
description = "Multi-agent architecture review system using AWS Bedrock"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = [
  "aws",
  "architecture",
  "review",
  "bedrock",
  "cloudformation",
  "ai",
  "agents",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: System :: Systems Administration",
]
dependencies = [
  "bedrock-agentcore>=1.1.1",
  "bedrock-agentcore-starter-toolkit>=0.1.0",
  "boto3>=1.42.2",
  "click>=8.0.0",
  "markdown>=3.10",
  "pillow>=12.0.0",
  "python-frontmatter>=1.1.0",
  "strands-agents>=1.19.0",
]

[project.optional-dependencies]
otel = ["strands-agents[otel]>=1.19.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["arch_sparring_agent"]

[dependency-groups]
dev = ["ruff>=0.14.8"]

[project.scripts]
arch-review = "arch_sparring_agent.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/arch-sparring-agent"
Documentation = "https://github.com/yourusername/arch-sparring-agent#readme"
Repository = "https://github.com/yourusername/arch-sparring-agent"
Changelog = "https://github.com/yourusername/arch-sparring-agent/blob/main/CHANGELOG.md"
Issues = "https://github.com/yourusername/arch-sparring-agent/issues"
